Use project version instead of generated version.
The reason we don't remove the generated version task is because
the build-and-upload task depends on a version resource as one
of its inputs. So instead we overwrite the generated version in
the local get-version task. We could write the version out to a
different location so it could be reused across tasks and you
could pick either output as input into your task that requires
a version. More to come on that soon...